requirements.txt
tensorflow-gpu~=2.6.0
pandas~=1.3.2
numpy~=1.19.5
matplotlib~=3.4.3
sklearn~=0.0
scikit-learn~=0.24.2
geopandas~=0.9.0
Shapely~=1.7.1
haversine~=2.3.1